% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{fit_model}
\alias{fit_model}
\title{fit a model to the data using a specified fitting method}
\usage{
fit_model(
  y,
  w,
  x,
  rotate_variables = FALSE,
  covariance_matrix = NULL,
  fitting_method = "VB",
  ...
)
}
\arguments{
\item{y}{Numeric vector}

\item{w}{Numeric matrix}

\item{x}{Numeric matrix}

\item{rotate_variables}{Logical value indicating whether to rotate y,w,x using covariance_matrix}

\item{covariance_matrix}{Numeric matrix used for rotation of y,w,x if NULL and rotate_variables is TRUE then default matrix is used}

\item{fitting_method}{Character string indicating the method used for fitting the data possible values are 'VB' 'Gibbs' 'Adaptive_Gibbs'}

\item{...}{arguments to pass down to internal methods.}
}
\value{
returns an object of class 'DPR_Model'
}
\description{
fit a model to the data using a specified fitting method
}
