% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.DPR_Model}
\alias{predict.DPR_Model}
\title{Use a DPR model to predict results from new data}
\usage{
\method{predict}{DPR_Model}(dpr_model, newdata)
}
\arguments{
\item{dpr_model}{an object of class DPR_Model}

\item{newdata}{Numeric matrix representing the input to the model}
}
\value{
returns Numeric vector of predictions
}
\description{
Use a DPR model to predict results from new data
}
