% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.DPR_Model}
\alias{predict.DPR_Model}
\title{Use a DPR model to predict results from new data}
\usage{
\method{predict}{DPR_Model}(dpr_model, newdata)
}
\arguments{
\item{dpr_model}{an object of class DPR_Model}

\item{newdata}{Numeric matrix representing the input to the model}
}
\value{
returns Numeric vector of predictions
}
\description{
Use a DPR model to predict results from new data
}
\examples{
print(getwd())
n <- 500
p <- 10775
x = readRDS("tests/testthat/data/in/x.rds")
y = readRDS("tests/testthat/data/in/y.rds")
w = readRDS("tests/testthat/data/in/w.rds")
dpr_model <- fit_model(y, w, x, fitting_method = "VB")
new_x <- matrix(rnorm(n = n * p, mean = 0, sd = 1), nrow = n, ncol = p)
new_y <- predict(dpr_model, new_x)
}
